+++ datagrowth/resources/http/generic.py
@@ -603,13 +603,6 @@ class HttpResource(Resource):
         super(HttpResource, self).__init__(*args, **kwargs)
 
     def clean(self):
-        # Legacy HttpResource instances may have a JSON string as request and head
-        # We parse that JSON to actual data here
-        if isinstance(self.request, str):
-            self.request = json.loads(self.request)
-        if isinstance(self.head, str):
-            self.head = json.loads(self.head)
-        # Actual cleaning implementation
         if self.request and not self.uri:
             uri_request = self.request_without_auth()
             self.uri = self.uri_from_url(uri_request.get("url"))